refactor to job
